+++ R/app_server.R
@@ -5,6 +5,9 @@
 #' @import shiny
 #' @noRd
 app_server <- function( input, output, session ) {
+  
   # List the first level callModules here
-
+  callModule(mod_atividade_economica_server, "atividade_economica_ui_1")
+  callModule(mod_atividade_economica_es_server, "atividade_economica_es_ui_1")
+  
 }