+++ R/app_ui.R
@@ -3,16 +3,39 @@
 #' @param request Internal parameter for `{shiny}`. 
 #'     DO NOT REMOVE.
 #' @import shiny
+#' @import shinydashboard
+#' @import shinydashboardPlus
 #' @noRd
+
+
 app_ui <- function(request) {
+  
+  data("series")
+  
   tagList(
     # Leave this function for adding external resources
     golem_add_external_resources(),
     # List the first level UI elements here 
-    fluidPage(
-      h1("econdashboard")
+    dashboardPagePlus(
+      dashboardHeaderPlus(title = "Econ Dashboard"),
+      dashboardSidebar(width = 150,
+                       sidebarMenu(
+                         menuItem("Conjuntura", tabName = "conjuntura", icon = icon("chart-bar")),
+                         menuItem("Mapas", tabName = "mapas", icon = icon("globe-americas"))
+                       )
+      ),
+      dashboardBody(
+        tabItems(
+          tabItem(tabName = "conjuntura",
+                  mod_titulo_ui("titulo_ui_1"),
+                  mod_atividade_economica_ui("atividade_economica_ui_1"),
+                  mod_atividade_economica_es_ui("atividade_economica_es_ui_1")
+          )
+        )
+      )
     )
   )
+  
 }
 
 #' Add external Resources to the Application