+++ NEWS.md
@@ -1,3 +1,12 @@
+# econdashboard (development version)
+
+* Adicionada módulo de Atividade Econômica BR
+  - PIB (variação anual)
+  - PIB (Índice Trimestral)
+  - IBC-Br
+  - Varejo e Serviços
+  - Exportação de bens
+
 # econdashboard 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.